Add CustomTextFromField
